auth_viz: split into modules
