+++ scripts/auth-viz/package.json
@@ -2,10 +2,14 @@
   "name": "testchain_scripts",
   "version": "1.0.0",
   "main": "index.js",
+  "license": "AGPL-3.0-only",
   "dependencies": {
     "dagre": "^0.8.2",
     "graphlib-dot": "^0.6.2",
     "mz": "^2.7.0",
     "web3": "^1.0.0-beta.36"
+  },
+  "scripts": {
+    "start": "node ./src/index.js"
   }
 }
